MiniChefV2 overview optimization
10x speed increase!
Relocates script from `util/miniChefOverview.js` to `minichef/overview.js`
